+++ autotest/t004_test.py
@@ -8,6 +8,10 @@ import flopy
 
 # define program data
 target = 'mp6'
+if sys.platform.lower() == 'win32':
+    target += '.exe'
+
+# get program dictionary
 prog_dict = pymake.usgs_program_data.get_target(target)
 
 # set up paths
@@ -18,7 +22,7 @@ if not os.path.exists(dstpth):
 mp6pth = os.path.join(dstpth, prog_dict.dirname)
 expth = os.path.join(mp6pth, 'example-run')
 
-exe_name = 'mp6'
+exe_name = target
 srcpth = os.path.join(mp6pth, prog_dict.srcdir)
 epth = os.path.abspath(os.path.join(dstpth, target))
 
@@ -38,46 +42,56 @@ def compile_code():
 
 
 def get_simfiles():
-    simfiles = [f for f in os.listdir(expth) if f.endswith('.mpsim')]
+    if os.path.exists(epth):
+        simfiles = [f for f in os.listdir(expth) if f.endswith('.mpsim')]
+    else:
+        simfiles = [None]
+
     return simfiles
 
 
 def run_modpath6(fn):
-    # rename a few files for linux
-    replace_files = ['example-6', 'example-7', 'example-8']
-    for rf in replace_files:
-        if rf in fn.lower():
-            fname1 = os.path.join(expth, '{}.locations'.format(rf))
-            fname2 = os.path.join(expth, '{}_mod.locations'.format(rf))
-            print('copy {} to {}'.format(os.path.basename(fname1),
-                                         os.path.basename(fname2)))
-            shutil.copy(fname1, fname2)
-            print('deleting {}'.format(os.path.basename(fname1)))
-            os.remove(fname1)
-            fname1 = os.path.join(expth, '{}.locations'.format(rf.upper()))
-            print('renmae {} to {}'.format(os.path.basename(fname2),
-                                           os.path.basename(fname1)))
-            os.rename(fname2, fname1)
+    if os.path.exists(epth):
+        # rename a few files for linux
+        replace_files = ['example-6', 'example-7', 'example-8']
+        for rf in replace_files:
+            if rf in fn.lower():
+                fname1 = os.path.join(expth, '{}.locations'.format(rf))
+                fname2 = os.path.join(expth, '{}_mod.locations'.format(rf))
+                print('copy {} to {}'.format(os.path.basename(fname1),
+                                             os.path.basename(fname2)))
+                shutil.copy(fname1, fname2)
+                print('deleting {}'.format(os.path.basename(fname1)))
+                os.remove(fname1)
+                fname1 = os.path.join(expth, '{}.locations'.format(rf.upper()))
+                print('renmae {} to {}'.format(os.path.basename(fname2),
+                                               os.path.basename(fname1)))
+                os.rename(fname2, fname1)
 
-    # run the model
-    print('running model...{}'.format(fn))
-    success, buff = flopy.run_model(epth, fn, model_ws=expth, silent=False)
-    assert success, 'could not run...{}'.format(os.path.basename(fn))
+        # run the model
+        print('running model...{}'.format(fn))
+        success, buff = flopy.run_model(epth, fn, model_ws=expth, silent=False)
+        if not success:
+            errmsg = 'could not run...{}'.format(os.path.basename(fn))
+    else:
+        success = False
+        errmsg = '{} does not exist'.format(epth)
+
+    assert success, errmsg
     return
 
 
 def clean_up():
     # clean up download directory
     print('Removing folder ' + mp6pth)
-    shutil.rmtree(mp6pth)
-
-    ext = ''
-    if sys.platform == 'win32':
-        ext = '.exe'
+    if os.path.isdir(mp6pth):
+        shutil.rmtree(mp6pth)
 
     # clean up executable
     print('Removing ' + target)
-    os.remove(epth + ext)
+    if os.path.isfile(epth):
+        os.remove(epth)
+
     return
 
 
@@ -88,17 +102,18 @@ def test_compile():
 
 def test_modpath6():
     simfiles = get_simfiles()
+
     for fn in simfiles:
         yield run_modpath6, fn
 
 
 def test_clean_up():
-    yield clean_up
+    clean_up()
 
 
 if __name__ == "__main__":
-    compile_code()
+    test_compile()
     simfiles = get_simfiles()
     for fn in simfiles:
         run_modpath6(fn)
-    clean_up()
+    test_clean_up()
