+++ autotest/t010_test.py
@@ -7,6 +7,10 @@ import pymake
 
 # define program data
 target = 'gridgen'
+if sys.platform.lower() == 'win32':
+    target += '.exe'
+
+# get program dictionary
 prog_dict = pymake.usgs_program_data.get_target(target)
 
 # set up paths
@@ -17,11 +21,16 @@ if not os.path.exists(dstpth):
 ver = prog_dict.version
 pth = os.path.join(dstpth, prog_dict.dirname)
 expth = os.path.join(pth, 'examples')
+exe_name = os.path.join(dstpth, target)
 
 
 def get_example_dirs():
-    exdirs = [o for o in os.listdir(expth)
-              if os.path.isdir(os.path.join(expth, o))]
+    prog = os.path.abspath(exe_name)
+    if os.path.isfile(prog):
+        exdirs = [o for o in os.listdir(expth)
+                  if os.path.isdir(os.path.join(expth, o))]
+    else:
+        exdirs = [None]
     return exdirs
 
 
@@ -36,20 +45,19 @@ def compile_code():
     pymake.build_program(target=target, fc=None, cc='g++',
                          include_subdirs=True,
                          download_dir=dstpth,
+                         exe_dir=dstpth,
                          replace_function=replace_function)
 
 
 def clean_up():
     # clean up
     print('Removing folder ' + pth)
-    shutil.rmtree(pth)
-
-    ext = ''
-    if sys.platform == 'win32':
-        ext = '.exe'
+    if os.path.isdir(pth):
+        shutil.rmtree(pth)
 
     print('Removing ' + target)
-    os.remove(target + ext)
+    if os.path.isfile(exe_name):
+        os.remove(exe_name)
     return
 
 
@@ -63,8 +71,6 @@ def run_command(cmdlist, cwd):
 
 
 def run_gridgen(d):
-    print('running...{}'.format(d))
-
     biscayne_cmds = [
         'buildqtg action01_buildqtg.dfn',
         'grid02qtg-to-usgdata action02_writeusgdata.dfn',
@@ -78,22 +84,28 @@ def run_gridgen(d):
         'grid02qtg-to-vtkfile action05_vtkfile.dfn',
         'grid02qtg-to-vtkfilesv action05_vtkfile.dfn', ]
 
-    testpth = os.path.join(expth, d)
-    testpth = os.path.abspath(testpth)
-    prog = os.path.abspath(target)
+    prog = os.path.abspath(exe_name)
+    if os.path.exists(prog):
+        print('running...{}'.format(d))
 
-    for cmd in biscayne_cmds:
-        cmdlist = [prog] + cmd.split()
-        print('running ', cmdlist)
-        retcode = run_command(cmdlist, testpth)
+        testpth = os.path.join(expth, d)
+        testpth = os.path.abspath(testpth)
+
+        for cmd in biscayne_cmds:
+            cmdlist = [prog] + cmd.split()
+            print('running {}'.format(' '.join(cmdlist)))
+            retcode = run_command(cmdlist, testpth)
+            success = False
+            if retcode == 0:
+                success = True
+            assert success, 'could not run {}'.format(' '.join(cmdlist))
+
+        if success:
+            pymake.teardown(testpth)
+    else:
         success = False
-        if retcode == 0:
-            success = True
-        assert success
 
-    if success:
-        pymake.teardown(testpth)
-    assert success is True
+    assert success, 'could not run {}'.format(prog)
 
     return
