+++ autotest/t012_test.py
@@ -7,6 +7,10 @@ import flopy
 
 # define program data
 target = 'gsflow'
+if sys.platform.lower() == 'win32':
+    target += '.exe'
+
+# get program dictionary
 prog_dict = pymake.usgs_program_data.get_target(target)
 
 # set up paths
@@ -29,50 +33,56 @@ exes = [egsflow]
 
 
 def copy_example_dir(epth):
-    src = os.path.join(expth, epth)
-    dst = os.path.join(dstpth, epth)
+    if os.path.exists(egsflow):
+        src = os.path.join(expth, epth)
+        dst = os.path.join(dstpth, epth)
 
-    # delete dst if it exists
-    if os.path.isdir(dst):
-        shutil.rmtree(dst)
+        # delete dst if it exists
+        if os.path.isdir(dst):
+            shutil.rmtree(dst)
 
-    # copy the files
-    try:
-        shutil.copytree(src, dst)
-    except:
-        msg = "could not move files from {} to '{}'".format(src, dst)
-        raise NameError(msg)
+        # copy the files
+        try:
+            shutil.copytree(src, dst)
+        except:
+            msg = "could not move files from {} to '{}'".format(src, dst)
+            raise NameError(msg)
 
-    # edit the control file for a shorter run
-    # sagehen
-    if epth == examples[0]:
-        fpth = os.path.join(dstpth, epth, 'linux', control_files[0])
-        with open(fpth) as f:
-            lines = f.readlines()
-        with open(fpth, 'w') as f:
-            idx = 0
-            while idx < len(lines):
-                line = lines[idx]
-                if 'end_time' in line:
-                    line += '6\n1\n1981\n'
-                    idx += 3
-                f.write(line)
-                idx += 1
+        # edit the control file for a shorter run
+        # sagehen
+        if epth == examples[0]:
+            fpth = os.path.join(dstpth, epth, 'linux', control_files[0])
+            with open(fpth) as f:
+                lines = f.readlines()
+            with open(fpth, 'w') as f:
+                idx = 0
+                while idx < len(lines):
+                    line = lines[idx]
+                    if 'end_time' in line:
+                        line += '6\n1\n1981\n'
+                        idx += 3
+                    f.write(line)
+                    idx += 1
     return
 
 
 def run_gsflow(example, control_file):
-    # copy files
-    copy_example_dir(example)
+    if os.path.exists(egsflow):
+        # copy files
+        copy_example_dir(example)
 
-    model_ws = os.path.join(dstpth, example, 'linux')
+        model_ws = os.path.join(dstpth, example, 'linux')
 
-    # run the flow model
-    msg = '{}'.format(egsflow)
-    success, buff = flopy.run_model(egsflow, control_file, model_ws=model_ws,
-                                    silent=False)
+        # run the flow model
+        success, buff = flopy.run_model(egsflow, control_file,
+                                        model_ws=model_ws, silent=False)
+        if not success:
+            errmsg = 'could not run {}'.format(control_file)
+    else:
+        success = False
+        errmsg = 'could not run...{}'.format(egsflow)
 
-    assert success, 'could not run...{}'.format(msg)
+    assert success, errmsg
 
     return
 
@@ -90,14 +100,10 @@ def clean_up():
             print('Removing example folder ' + example)
             shutil.rmtree(pth)
 
-    exe = egsflow
-    if sys.platform == 'win32':
-        exe += '.exe'
-
     # clean up compiled executables
-    if os.path.isfile(exe):
-        print('Removing ' + exe)
-        os.remove(exe)
+    if os.path.isfile(egsflow):
+        print('Removing ' + egsflow)
+        os.remove(egsflow)
     return
